pyproject.toml
[tool.poetry]
name = "petcarebutler"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.1"
pillow = "^12.1.0"
requests = "^2.32.5"
djangorestframework = "^3.16.1"
django-environ = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
